fix: fix passkey window
