+++ apps/ext/src/closePasskeyWIndow.ts
@@ -1,0 +1,16 @@
+import { EPassKeyWindowFrom } from '@onekeyhq/shared/src/utils/extUtils';
+
+const params = new URLSearchParams(globalThis.location.href.split('?').pop());
+const from = params.get('from') as EPassKeyWindowFrom;
+
+export const closeWindow = () => {
+  console.log('closeWindow');
+  if (
+    from === EPassKeyWindowFrom.popup ||
+    from === EPassKeyWindowFrom.sidebar
+  ) {
+    setTimeout(() => {
+      window.close();
+    }, 50);
+  }
+};
